+++ CHANGELOG.md
@@ -1,11 +1,13 @@
 Changelog
 =========
 
+  * New maintainer: Jiří Janoušek.
+
 1.2 - February 12th, 2017
 -------------------------
 
   * Ported to Nuvola SDK.
-  
+
 1.1 - October 22nd, 2016
 ------------------------